Moved Retry to git package
